Migrating the consensus tracker from TorCtl to Stem
The consensus tracker was a hack job. It does the trick in terms of notifying
us of basic Sybil attacks on the network, but it's only meant as a short term
solution until we invest time into a proper monitor and alarm framework.

Migrating the script from TorCtl to Stem. This is a pretty direct migration -
the behavior is unchanged and I'm not trying to clean up the script (though
stem does make it a bit more readable).

Doing this migration revealed a catastrophic performance bug in stem that is
now fixed...

https://gitweb.torproject.org/stem.git/commitdiff/bbd702c81e1923bfbab236d0d3649fd2f87b3d95

The memory usage is still a fair bit higher (around 5.1% of the memory on my
netbook verses 1.9% before). However, this isn't entirely unexpected. In a lot
of cases TorCtl really didn't do much, providing us raw strings rather than
parsed objects. Still, there's certainly plenty of low hanging fruit we should
address in stem to lower that gap.