Replacing consensusTracker.py with a simple sybil checker
Our consensusTracker.py was petty slick, but most of what it does is
unnecessary. The html formatted emails were pretty, and the stats were kinda
interesting. But the only true usefulness of the script was to notify us of
sybil attackers spamming us with relays. Replacing it with a far smaller,
simpler script that just does that.