Clearing last_notified prior to loading
Presently our last_notified.cfg is accumilating duplicate entries. For instance...

  bad_exits_in_sync 1376949957
  bad_exits_in_sync 1376949957
  bad_exits_in_sync 1376949957
  bandwidth_authorities_in_sync 1376952152
  cert_expiration.two_months.urras 1376795678
  cert_expiration.two_months.urras 1376795678
  ... etc...

This is a little weird since we include set()'s overwrite argument which should
avoid this. I suspect what's happening is that our load() is duplicating the
values since we're calling it for every suppression check (get_config() is a
singleton getter, so this is calling load on the same object). Clearing our
contents to see if it addresses this.