Sending descriptor_checker.py notices to me rather than the list
On reflection descriptor_checker.py provides us an alert of only two things:

1. An inability by stem to parse descriptors. This could be either due to
   malformed descriptor content or an issue with stem.

2. Directory authority outages.

The first I'd rather not send to tor-consensus-health@, and the second is
already provided by consensus_health_checker.py (though this script's notices
are a little nicer).

Hence I'm gonna keep this around as-is, except with notices going to me rather
than the list.