Suppressing sybil notification if not run in the last three hours
Our sybil checker compares a list of all fingerprints we've ever seen to the
current consensus to see how many new additions there have been. As such it
always goes off if it hasn't been run for a long while.

We got one of these false alarms when yatei went down for thirty hours.
Suppressing email notifications when we haven't run in the last three hours.