Dropping Tonga from checks
We were downloading Tonga's consensus but not its vote so it would be included
in checks for the former but not later. Up side is that this checked that Tonga
was up and healthy (still important since it's an authority, even it it doesn't
vote). But this is already covered by 'descriptor_checker.py'.

This is in response to...

https://lists.torproject.org/pipermail/tor-dev/2013-October/005618.html