Dropping parameter mismatch check
Authoity parameters have been out of sync for a while now and Roger reports
that it's intentional. Two authoities are setting an NumDirectoryGuards and
NumEntryGuards. Though it's not making it into the consensus they're doing this
so no single other authority can make it something different.

Dropping this notification since it's unactionable.