Messages about unmesured relays weren't being effectively suppressed
We suppress messages based on their content but the notice about unmeasured
relays have stats that change each hour. As a result they effectively weren't
being suppressed causing a notice each hour.

Adding a hack to clobber the dynamic data so our suppression string will
work...

  As_a_bandwidth_authority_gabelmoo_lacked_a_measurement_for_0_of_0_relays_(0%) 1409527674

I'm not thrilled with this solution, but we can add generic handling when we
have more than one message that needs it.