Properly suppress staggered messages
We updated suppression times if and only if its duration had expired. Take the
following scenario with the suppression time of six hours per message...

  2:00 - issue A comes up
  4:00 - issue B comes up

At 4:00 an email is sent having both issue A and B, so the proper thing would
be to suppress both until 10:00. However, message A's suppression time wasn't
being updated so we'd email at both 8:00 and 10:00 instead.

This should greatly cut down on the amount of noise since all messages have a
default suppression time.