Free memory prior to sending email
Frequently we run into OOM issues when sending a notification email at the end
of our run...

  Traceback (most recent call last):
    File "/srv/doctor.torproject.org/doctor/consensus_health_checker.py", line 705, in <module>
      util.send("Script Error", body_text = msg, destination = util.ERROR_ADDRESS)
    File "/srv/doctor.torproject.org/doctor/util.py", line 97, in send
      stderr = subprocess.PIPE,
    File "/usr/lib/python2.7/subprocess.py", line 679, in __init__
      errread, errwrite)
    File "/usr/lib/python2.7/subprocess.py", line 1153, in _execute_child
      self.pid = os.fork()
  OSError: [Errno 12] Cannot allocate memory

This is because we call subprocess.Popen which forks our process, doubling the
amount of memory we use. Sending an email obviously doesn't require much, but
since we're using hundreds of megs for all the consensuses we download that's
what is allocated.

  http://stackoverflow.com/a/13329386/1067192

Attempting to address this by freeing descriptor content prior to forking our
process.