Shift bot notification to generate subprocess early too
When I made the prior commit I knew there were a couple other spots we forked,
but wanted to see if it worked for the main email first. Seems to, so shifting
the bot notice too to prevent OOMs. The error notification still forks, but
leaving it as-is for the moment since it seems to be working (that's how I'm
being notified of OOMs, after all).