Don't shell out to send emails
I'm glad our recent DocTor change forced me to revert my earlier hack to get
around our OOM errors. On reflection it was stupid.

We shelled out to 'mail' because that's what the java DocTor implementation
did, and I wasn't really sure how to interact with the local mail daemon. But
that's dump - it's simply a local SMTP daemon so we can simply use the builtin
module.

This avoids shelling out, and by extension both gets rid of an ugly hack and
avoids forking our process (which triggered the OOM errors).