Script to check for fingerprint changes
During our recent dev meeting David Goulet asked me for a DocTor check for
relays rapidly changing their fingerprint. Doing so likely means they're trying
to get into a privilaged position to snoop on hidden services.

Plan is to keep an eye on this script for a bit (just whipped it up, no doubt
it has issues), then send the results to David when we have some confidence in
it. Later the results will be sent to bad-relays@.