Report fallback directory DirPort failures
The whole point of our fallback directory daemon is to tell us how many
fallbacks are unusable, but now that one of their DirPorts is misbehaving
it's resulting in uncaught exceptions. Interestingly, we're getting quite a
range...

  Traceback (most recent call last):
    File "fallback_directories.py", line 82, in <module>
      main()
    File "fallback_directories.py", line 57, in main
      downloader.get_consensus(endpoints = [(relay.address, relay.dir_port)]).run()
    File "/home/atagar/Desktop/tor/doctor/stem/descriptor/remote.py", line 445, in run
      return list(self._run(suppress))
    File "/home/atagar/Desktop/tor/doctor/stem/descriptor/remote.py", line 456, in _run
      raise self.error
  HTTPError: HTTP Error 404: Not Found

... as well as...

  HTTPError: HTTP Error 503: Directory busy, try again later

... or...

  CertificateError: hostname '85.214.62.48' doesn't match either of 'andamur.com', 'www.andamur.com'

Our DescriptorDownloader raises an unpleasantly large array of exceptions so
unfortunately requires a catch-all here (maybe I'll change that at some point
in stem...).