Bump fallback notification theshold (25% => 50%)
This has been notifying everyone for weeks due to breaching the 25% thresold.
Four days ago I reached out to teor and Nick to see what we want to do with
this but haven't heard back.

DocTor notices need to spur action or they're just noise. Bumping the threshold
to 50% so this no longer emails us.